requirements.txt
selene==1.0.2
pytest==4.6.10
selenium==3.141.0
uuid==1.30
pytest-testdox==2.0.1
tenacity==6.3.1
pytest-env==0.6.2
pytest-rerunfailures==8.0
faker==6.0.0
allure-pytest==2.8.33
pytest-xdist==1.34.0